requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
mysql-connector-python==8.2.0
pymysql==1.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
itsdangerous==2.1.2
reportlab==4.0.7
